rewrote testing for kcc state machine test base
