+++ Packages/com.nickmaltbie.openkcc/OpenKCC/AssemblyInfo.cs
@@ -20,3 +20,4 @@ using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("nickmaltbie.OpenKCC.Editor")]
 [assembly: InternalsVisibleTo("nickmaltbie.OpenKCC.Tests.EditMode")]
+[assembly: InternalsVisibleTo("nickmaltbie.OpenKCC.Tests.TestCommon")]
