+++ Packages/com.nickmaltbie.openkcc/Tests/EditMode/Character/Config/KCCGroundedStateTests.cs
@@ -18,7 +18,7 @@
 
 using Moq;
 using nickmaltbie.OpenKCC.Character.Config;
-using nickmaltbie.OpenKCC.TestCommon;
+using nickmaltbie.OpenKCC.Tests.TestCommon;
 using nickmaltbie.OpenKCC.Utils;
 using nickmaltbie.TestUtilsUnity.Tests.TestCommon;
 using NUnit.Framework;
