+++ Packages/com.nickmaltbie.openkcc/Tests/EditMode/Utils/KCCUtilsTests.cs
@@ -20,7 +20,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Moq;
 using nickmaltbie.OpenKCC.Character;
-using nickmaltbie.OpenKCC.TestCommon;
+using nickmaltbie.OpenKCC.Tests.TestCommon;
 using nickmaltbie.OpenKCC.Utils;
 using nickmaltbie.TestUtilsUnity.Tests.TestCommon;
 using NUnit.Framework;
