+++ Packages/com.nickmaltbie.openkcc/Tests/PlayMode/Utils/KCCUtilsScenarioTests.cs
@@ -22,7 +22,7 @@ using System.Collections.Generic;
 using System.Linq;
 using nickmaltbie.OpenKCC.Character;
 using nickmaltbie.OpenKCC.Environment.Pushable;
-using nickmaltbie.OpenKCC.TestCommon;
+using nickmaltbie.OpenKCC.Tests.TestCommon;
 using nickmaltbie.OpenKCC.Utils;
 using nickmaltbie.TestUtilsUnity.Tests.TestCommon;
 using NUnit.Framework;
