+++ Packages/com.nickmaltbie.openkcc/Tests/PlayMode/nickmaltbie.OpenKCC.Tests.PlayMode.asmdef
@@ -2,18 +2,18 @@
     "name": "nickmaltbie.OpenKCC.Tests.PlayMode",
     "rootNamespace": "nickmaltbie.OpenKCC.Tests.PlayMode",
     "references": [
+        "nickmaltbie.OpenKCC",
+        "nickmaltbie.OpenKCC.Tests.TestCommon",
+        "nickmaltbie.StateMachineUnity",
+        "nickmaltbie.TestUtilsUnity",
+        "nickmaltbie.TestUtilsUnity.Tests.TestCommon",
         "Unity.InputSystem",
         "Unity.InputSystem.TestFramework",
         "Unity.ProBuilder",
         "Unity.ProBuilder.Editor",
         "Unity.ProBuilder.Poly2Tri",
-        "nickmaltbie.OpenKCC",
-        "nickmaltbie.OpenKCC.TestCommon",
         "UnityEngine.TestRunner",
-        "nickmaltbie.StateMachineUnity",
-        "UnityEditor.TestRunner",
-        "nickmaltbie.TestUtilsUnity",
-        "nickmaltbie.TestUtilsUnity.Tests.TestCommon"
+        "UnityEditor.TestRunner"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
