+++ Packages/com.nickmaltbie.openkcc/Tests/TestCommon/KCCStateMachineTestBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: da363ad2abeecd340bf3266b51556328
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
