+++ Packages/com.nickmaltbie.openkcc/Tests/TestCommon/nickmaltbie.OpenKCC.Tests.TestCommon.asmdef.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9f3325cc3a6bb394f940c97b0dc48a39
+guid: c1110867878f805408a1526001257b83
 AssemblyDefinitionImporter:
   externalObjects: {}
   userData:
